Refactorización del componente FlashCardGenerator
