+++ src/components/generator/FlashCardGenerator.tsx
@@ -1,4 +1,5 @@
 
+import { FormEvent } from "react"
 import ThemeSelectorComponent from "./ThemeSelector";
 import { Button } from "../ui/button";
 import { Textarea } from "../ui/textarea";
@@ -6,13 +7,15 @@ import { Send } from "lucide-react";
 import type { UserResource } from "@clerk/types";
 import { SourceSelector } from "./SourceSelector";
 import { InfoCards } from "../cards/info-cards";
-import { useState } from "react";
+import { useThemeStore } from "@/store/interestThemes";
+import { processQuestions } from "@/utils/services/functions/process/processQuestion";
+import { useGetAnswers } from "@/hooks/useGetAnswers";
+import { useFlashCardsStore } from "@/store/flashCardsStore";
 
 interface Props {
   error: string | null;
   pregunta: string;
   loadingAnswers: boolean;
-  handleSubmit: (e: React.FormEvent<HTMLFormElement>) => void;
   handlePreguntaChange: (e: React.ChangeEvent<HTMLTextAreaElement>) => void;
   user: UserResource | null | undefined
   textAreaRef: React.RefObject<HTMLTextAreaElement | null>;
@@ -24,12 +27,56 @@ export const FlashCardGenerator: React.FC<Props> = ({
   error,
   pregunta,
   loadingAnswers,
-  handleSubmit,
   handlePreguntaChange,
   user,
   textAreaRef
 }) => {
-  const [selectedSource, setSelectedSource] = useState("all");
+  const addNewFlashcards = useFlashCardsStore((state) => state.addNewFlashcards);
+  const selectedTheme = useThemeStore((state) => state.selectedTheme);
+  const {  getAnswers } = useGetAnswers()
+  
+  const handleSubmit = async (e: FormEvent<HTMLFormElement>) => {
+    e.preventDefault();
+    const formData = new FormData(e.currentTarget);
+    const data = Object.fromEntries(formData.entries());
+    const result = processQuestions({data})
+
+    if(result.error) {
+      setError(result.error);
+      setTimeout(() => {
+        setError(null);
+      }, 2000); 
+      return;
+    }
+      const questions = result.questions as string[];
+
+    try {
+      // Obtenemos el tema
+      const theme = selectedTheme;
+      
+      //Enviamos los datos a la API
+      const userLevel = "basic";
+      
+      // La respuesta ya es el array de respuestas directamente
+      const answers = await getAnswers({ theme, questions, userLevel });
+      
+      if (!answers) {
+        throw new Error("No se recibieron respuestas de la API");
+      }
+
+      //Enviamos los datos al store para mostrar de forma local
+      addNewFlashcards(theme, questions, answers);
+
+      //resetForm();
+    } catch (error) {
+      setError(
+        error instanceof Error
+          ? error.message
+          : "Error al procesar la solicitud"
+      );
+    }
+  };
+
 
   return (
     <div className="flex flex-col items-center justify-center min-h-screen w-full  p-4 md:p-8">
@@ -42,18 +89,13 @@ export const FlashCardGenerator: React.FC<Props> = ({
           <p className="text-sm md:text-base text-[#84aed5] mb-2 text-center">
             POTENCIADO POR IA PARA MEJORAR TU APRENDIZAJE
           </p>
-        </div>
-        {/* Selector de tema separado visualmente */}
-        <div className="w-full flex justify-center">
+        </div>        
             <ThemeSelectorComponent />
-        </div>
-        {/* Formulario principal */}
         <form
           onSubmit={handleSubmit}
           className="w-full bg-opacity-[0.03] border border-[#4a525a]/20 rounded-xl shadow-lg"
         >
-          {/* Textarea separado visualmente */}
-          <div className="p-4 md:p-6 flex flex-col gap-4">
+          <section className="p-4 md:p-6 flex flex-col gap-4">
             <Textarea
               ref={textAreaRef}
               id="pregunta"
@@ -68,9 +110,9 @@ export const FlashCardGenerator: React.FC<Props> = ({
               aria-label="Introduce un tema o concepto para tus flashcards"
               required
             />
-          </div>
-          <div className="border-t border-[#4a525a]/20 p-3 flex flex-col md:flex-row items-center justify-between gap-3">
-            <SourceSelector selected={selectedSource} onSelect={setSelectedSource} />
+          </section>
+          <section className="border-t border-[#4a525a]/20 p-3 flex flex-col md:flex-row items-center justify-between gap-3">
+            <SourceSelector />
             <Button
               type="submit"
               disabled={!user || !user.id || !!error || loadingAnswers}
@@ -81,14 +123,13 @@ export const FlashCardGenerator: React.FC<Props> = ({
               <Send className="w-4 h-4 mr-2" />
               Generar Flashcards
             </Button>
-          </div>
-          {error && (
+          </section>
+        </form>
+        {error && (
             <p className="text-red-500 text-center px-4 pb-2" aria-live="assertive">
               {error}
             </p>
           )}
-        </form>
-        {/* Tarjetas informativas */}
        <InfoCards />
       </div>
     </div>
